Final release
Functions have been tested and debugged.

Bounce function added. It is called internally by getpkt when IDS is not
the same as the local ID and transmits the message to the proper IDS,
then gets the response back to the original requester.

Simple rain/window code implemented for testing purposes. Can be found
under the sensorAgua and Janela files.